It doesn't make sense for @Option to have multi-value setting.
This is because whether a setter can handle multiple invocations
well (by retaining all the values) or not (usually by having the last
invocation overwrites all the previous values) are determined
by Setter and setter alone, so asking the user to configure that makes
no sense.

MultivaluedTest indicates that the author of this part of the code seems
to have been thinking about concatanating values or something like that,
but if someone wants that kind of unusual parsing, he should be writing
OptionHandler.