Fixed the way MapOptionHandler works
The problem with MapSetter is that it has too much logic in it. Setters are not meant to be doing too much --- it's merely there to abstract away the difference between fields and methods. Anything that involves parsing should be OptionHandler.

The reason MapOptionHandler wasn't done cleanly is because setting a Map requires a get access of the field value. So we should expose this from FieldSetter.