Massaged the contribution from @andrewgaul.
The hidden() switch that always hides it from the usage screen is pointless, as it can be trivially achieved by not writing any usage. I think the underlying issue the change was trying to address is that being able to show different subset of options is often useful. Such a use case can be already seen in ExampleMode.

In this change, I introduced a generalized version of ExampleMode called OptionHandlerFilter, which allows the caller to specify the subset of options. The printUsage method is extended to support this parameter, and in this way the caller can control the subset of options to print them.