Renamed the files to match the ResourceBundle file lookup logic and fix a small bug.
The file Messages_de_DE.properties will ONLY be found for people speaking German and being in Germany. Austrians or Switzerland citizens also speak German,
but for them this file won't be found because of the ResourceBundle resolution logic.

The file name format for ResourceBundle property files can be:

     baseName + "_" + language + "_" + country

or
      baseName + "_" + language

(other options omitted)

The resolution algorithm is explained in the getBundle() Javadoc here:

	http://docs.oracle.com/javase/7/docs/api/java/util/ResourceBundle.html

The algorithm will start with the most specific locale matching file name variant and iteratively strip off parts.
It will never try other countries that seem 'friendly' or 'compatible'.

See this sentence for a clue:

 "If there are different resources for different countries, you can make specializations: for example, "MyResources_de_CH" contains objects for the German language (de) in Switzerland (CH). If you want to only modify some of the resources in the specialization, you can do so. "

This could be used for example country specific specialities which are probably not relevant for args4j.