Disable @-syntax by default
The @-syntax allows any user passing options parsed by args4j to
effectively read any file on the machine running the args4j application.
For the common use case of building a command-line application that can
be run by a user with shell access to a machine, this is a useful
shortcut. It doesn't increase the surface of files that the user has
access to; if they already have shell access to a machine, they can read
any files they want with `cat`.

However, args4j may also be used in environments where command-line
parsing is accessible to arbitrary users over a network. For example,
args4j can be used in conjunction with an SSH server library such as
Apache MINA to expose a limited set of SSH commands without granting
shell access to the user. In this context, users should definitely not
be allowed to read arbitrary files from the server using @-syntax.

Because the behavior is non-obvious to an implementer who doesn't
closely read the ParserProperties documentation, and because the impact
of exposing arbitrary files on a remote server is potentially so severe,
change the default for @-syntax to off.